requirements.txt
python-telegram-bot==20.*
openai>=1.0.0
python-dotenv
Flask
requests